Split Generator into an engine trait and HTML impl
Generic behaviour, with regards to the AST engine types, has been moved
onto default methods on a Generator trait in the engine module, which
has associated types and method declarations to be filled in by an
implementation.
The HTML implementation of the engine Generator trait should function
the same as it did before, as the only thing that has happened is code
moving.